Refactor AttachmentView.swift
* Move attachment views to their own files
* Move some functions in AttachmentView to an extension
* Add some SwiftUI previews for certain types of attachments (they don’t work yet due to some Sparkle issue)

Also updated ServerButtonStyle to use new extensions
